Support admin-integrated provisioning flow
Signed-off-by: Aayush Gupta <aayushgupta219@gmail.com>
Change-Id: I3f0273af1b28dfd873bea20f9b8605850f671c9c

Start Orbot automatically for Safer and Safest

Orbot 16.6.4-RC-1+ needs to be started just once, and then it will
start automatically on future boots. After profile provisioning, start
it in the background using a targeted broadcast, which the app allows
by default.

Co-authored-by: Aayush Gupta <aayushgupta219@gmail.com>
Issue: calyxos#1523
Change-Id: I303ebfedb295dd90983821c0de348133a58bb0a7

Allow notifications for garlic-provisioned Orbot

This allows the user to know the status of Orbot even if they have
never opened it manually.

Change-Id: I6f05ca9b3fee6beedfc4ba7d53da4ab8410f05b4