Support admin-integrated provisioning flow
Orbot 16.6.4-RC-1+ needs to be started just once, and then it will
start automatically on future boots. After profile provisioning, start
it in the background using a targeted broadcast, which the app allows
by default.

Also allow notifications for garlic-provisioned Orbot. This allows
the user to know the status of Orbot even if they have never opened
it manually.

Change-Id: I3f0273af1b28dfd873bea20f9b8605850f671c9c