Prevent Bellis from launching SetupWizard twice
When work profile provisioning completes, MainActivity of Bellis is
launched, and in onCreate, it responds by launching the Setup Wizard
for the work profile. When that wizard completes, we are returned to
Bellis, where onCreate is called again, and the starting intent is
the same as the first time.

Store a value in our savedInstanceState to confirm that we have already
handled the necessary tasks in onCreate, so that we don't do it twice.

Issue: calyxos#1703
Change-Id: Id94b9b8696df36f07dc754abb57f6b5abbc8e57f