Return result before launching Setup Wizard
The goal of this change is to send the successful result to
ManagedProvisioning immediately so that even if the Work Profile
Setup Wizard is not completed, the provisioning state for Safest
(or Safer) will be properly set. Otherwise, the provisioning state
is not properly set until after the Work Profile Setup Wizard
successfully completes, which is not guaranteed.

A side effect of an improper provisioning state is that additional
work profiles can never again be added.

Issue: calyxos#2022
Change-Id: I75b139021b9f55bf5bcb8f799b05b998f3c19449